Sistema de Login, Logout e Cadastro

+++ estoque/views.py
@@ -4,10 +4,16 @@ from .forms import EstoqueForm, ProdutoForms, EditarEstoqueForm, QuantidadeForm
 from django.contrib import messages
 
 def mostrar_estoque(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     estoque = Estoque.objects.all()
-    return render(request, 'index.html', {'estoque': estoque})
+    return render(request, 'estoque/index.html', {'estoque': estoque})
 
 def adicionar_ao_estoque(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     if request.method == 'POST':
         form = EstoqueForm(request.POST)
         if form.is_valid():
@@ -25,7 +31,7 @@ def adicionar_ao_estoque(request):
             produto_existente = Estoque.objects.filter(produto=form.cleaned_data['produto']).first()
             if produto_existente:
                 messages.error(request, f'Produto "{form.cleaned_data['produto']}" já existe no estoque!')
-                return redirect('mostrar_estoque')
+                return redirect('estoque/mostrar_estoque')
             else:
                 novo_produto = form.save(commit=False)
                 novo_produto.fornecedor = fornecedor
@@ -33,17 +39,20 @@ def adicionar_ao_estoque(request):
                 novo_produto.preco = preco
                 novo_produto.save()
                 messages.success(request, f'Produto "{form.cleaned_data['produto']}" adicionado ao estoque com sucesso!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
         else:
             messages.error(request, 'Erro ao adicionar produto!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
         
     else:
         form = EstoqueForm()
     produtos = Produto.objects.all()
-    return render(request, 'adicionar_ao_estoque.html', {'form': form, 'produtos': produtos})
+    return render(request, 'estoque/adicionar_ao_estoque.html', {'form': form, 'produtos': produtos})
 
 def editar_produto(request, id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     produto = get_object_or_404(Estoque, id=id)
     if request.method == 'POST':
         form = EditarEstoqueForm(request.POST, instance=produto)
@@ -64,33 +73,42 @@ def editar_produto(request, id):
             produto.preco = preco
             produto.save()
             messages.success(request, f'Produto "{produto.produto}" editado com sucesso!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
         else:
             messages.error(request, 'Erro ao editar produto!')
     else:
         form = EditarEstoqueForm(instance=produto)
-    return render(request, 'editar_produto.html', {'form': form, 'produto': produto})
+    return render(request, 'estoque/editar_produto.html', {'form': form, 'produto': produto})
 
 def excluir_produto(request,id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     produto = get_object_or_404(Estoque, id=id)
     produto.delete()
     messages.success(request, f'Produto "{produto.produto}" excluído com sucesso!')
-    return redirect('mostrar_estoque')
+    return redirect('estoque/mostrar_estoque')
 
 def adicionar_novo_produto(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     if request.method == 'POST':
         form = ProdutoForms(request.POST)
         if form.is_valid():
             form.save()
-            messages.success(request, f'Produto "{form.cleaned_data['nome']} adicionado com sucesso!')
-            return redirect('mostrar_estoque')
+            messages.success(request, f'Produto "{form.cleaned_data['nome']}" adicionado com sucesso!')
+            return redirect('estoque/mostrar_estoque')
         else:
             messages.error(request, 'Erro ao adicionar produto!')
     else:
         form = ProdutoForms()
-    return render(request, 'adicionar_novo_produto.html', {'form':form})
+    return render(request, 'estoque/adicionar_novo_produto.html', {'form':form})
 
 def aumentar_quantidade(request, id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     produto = get_object_or_404(Estoque, id=id)
     if request.method == 'POST':
         form = QuantidadeForm(request.POST)
@@ -98,15 +116,18 @@ def aumentar_quantidade(request, id):
             produto.quantidade += form.cleaned_data['quantidade']
             produto.save()
             messages.success(request, f'Quantidade de "{produto.produto}" aumentada com sucesso!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
         else:
             messages.error(request, 'Erro ao aumentar quantidade!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
     else:
         form = QuantidadeForm(initial={'quantidade': produto.quantidade})
-    return render(request, 'aumentar_quantidade.html', {'form': form, 'produto': produto})
+    return render(request, 'estoque/aumentar_quantidade.html', {'form': form, 'produto': produto})
 
 def diminuir_quantidade(request,id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     produto = get_object_or_404(Estoque, id=id)
     if request.method == 'POST':
         form = QuantidadeForm(request.POST)
@@ -114,18 +135,21 @@ def diminuir_quantidade(request,id):
             produto.quantidade -= form.cleaned_data['quantidade']
             produto.save()
             messages.success(request, f'Quantidade de "{produto.produto}" diminuida com sucesso!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
         else:
             messages.error(request, 'Erro ao diminuir quantidade!')
-            return redirect('mostrar_estoque')
+            return redirect('estoque/mostrar_estoque')
     else:
         form = QuantidadeForm(initial={'quantidade': produto.quantidade})
-    return render(request, 'diminuir_quantidade.html', {'form': form, 'produto': produto})
+    return render(request, 'estoque/diminuir_quantidade.html', {'form': form, 'produto': produto})
 
 def buscar_produto(request):
+    if not request.user.is_authenticated:
+        return redirect('login')
+    
     query = request.GET.get('buscar', '')
     if query:
         produtos = Estoque.objects.filter(produto__nome__icontains=query)
     else:
         produtos = Estoque.objects.all()
-    return render(request, 'buscar_produto.html', {'estoque': produtos})
+    return render(request, 'estoque/buscar_produto.html', {'estoque': produtos})
